testing how to create a cuda library

+++ software/em/xmipp/libraries/reconstruction_cuda/cuda_gpu_rotate_image.h
@@ -1,2 +1,2 @@
 
-void cuda_rotate_image(float *image, float *rotated_image, int ang);
+void cuda_rotate_image(float *image, float *rotated_image, float ang);
